backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and Supabase
supabase==2.3.0
asyncpg==0.29.0
sqlalchemy==2.0.23

# AI and Voice Processing
openai==1.3.0
elevenlabs==0.2.26
whisper-openai==20231117

# HTTP and WebSocket
httpx==0.25.2
websockets==12.0
python-multipart==0.0.6

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Email and Communication
jinja2==3.1.2
aiosmtplib==3.0.1

# File handling and storage
aiofiles==23.2.1
pillow==10.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Production
gunicorn==21.2.0 